+++ pydv/docs/release_notes.rst
@@ -10,6 +10,7 @@ If there is a domain specified, these functions will only return points in that
 domain.
 
 
+
 3.0.7
 -----
 * Fixed a bug in PyDV internal when using min and max over a range.